+++ platform/plugins/system/AP.Plugin.SystemSettings/Editors/PlcConfigurationEditorView.xaml.cs
@@ -1,0 +1,14 @@
+using System.Windows.Controls;
+
+namespace AP.Plugin.SystemSettings.Editors;
+
+/// <summary>
+/// PlcConfigurationEditorView.xaml 的交互逻辑
+/// </summary>
+public partial class PlcConfigurationEditorView : UserControl
+{
+    public PlcConfigurationEditorView()
+    {
+        InitializeComponent();
+    }
+}